remove utils module
