+++ examples/build.gradle
@@ -86,7 +86,6 @@ dependencies {
     implementation project(':libdirections-offboard')
     implementation project(':libdirections-onboard')
     implementation project(':libtrip-notification')
-    implementation project(':libnavigation-util')
     implementation project(':liblogger')
     implementation project(':libnavigator')
