+++ libdirections-offboard/build.gradle
@@ -19,7 +19,6 @@ dependencies {
     compileOnly dependenciesList.mapboxAnnotations
     kapt dependenciesList.mapboxAnnotationsProcessor
 
-    implementation project(':libnavigation-util')
     api project(':libnavigation-base')
 
     // Support
