+++ libdirections-offboard/src/test/java/com/mapbox/navigation/route/offboard/router/NavigationRouteEventListenerTest.kt
@@ -1,6 +1,6 @@
 package com.mapbox.navigation.route.offboard.router
 
-import com.mapbox.navigation.utils.time.ElapsedTime
+import com.mapbox.navigation.base.utils.time.ElapsedTime
 import io.mockk.mockk
 import io.mockk.verify
 import okhttp3.Call
