+++ libdirections-onboard/build.gradle
@@ -21,7 +21,6 @@ dependencies {
 
     api project(':libnavigation-base')
     implementation project(':libnavigator')
-    implementation project(":libnavigation-util")
 
     //ktlint
     ktlint dependenciesList.ktlint
