+++ libnavigation-base/src/main/java/com/mapbox/navigation/base/utils/extensions/NullEx.kt
@@ -1,4 +1,4 @@
-package com.mapbox.navigation.utils.extensions
+package com.mapbox.navigation.base.utils.extensions
 
 inline fun <R1, T> ifNonNull(r1: R1?, func: (R1) -> T): T? =
     if (r1 != null) {
