+++ libnavigation-base/src/main/java/com/mapbox/navigation/base/utils/span/SpanItem.kt
@@ -1,0 +1,6 @@
+package com.mapbox.navigation.base.utils.span
+
+interface SpanItem {
+
+    val span: Any
+}
