+++ libnavigation-base/src/main/java/com/mapbox/navigation/base/utils/span/TextSpanItem.kt
@@ -1,3 +1,3 @@
-package com.mapbox.navigation.utils.span
+package com.mapbox.navigation.base.utils.span
 
 class TextSpanItem(override val span: Any, val spanText: String) : SpanItem
