+++ libnavigation-metrics/build.gradle
@@ -20,7 +20,6 @@ android {
 
 dependencies {
     api project(':libnavigation-base')
-    implementation project(':libnavigation-util')
     api (dependenciesList.mapboxEvents) {
         exclude group: "com.mapbox.mapboxsdk", module: "mapbox-android-core"
     }
