+++ libnavigation-metrics/src/test/java/com/mapbox/navigation/metrics/MapboxMetricsReporterTest.kt
@@ -6,8 +6,8 @@ import com.mapbox.android.telemetry.Event
 import com.mapbox.android.telemetry.MapboxTelemetry
 import com.mapbox.navigation.base.metrics.MetricEvent
 import com.mapbox.navigation.base.metrics.NavigationMetrics
+import com.mapbox.navigation.base.utils.thread.WorkThreadHandler
 import com.mapbox.navigation.metrics.internal.utils.extensions.toTelemetryEvent
-import com.mapbox.navigation.utils.thread.WorkThreadHandler
 import io.mockk.mockk
 import io.mockk.verify
 import org.junit.Test
