--- libnavigation-util/build.gradle
@@ -1,31 +1,0 @@
-apply plugin: 'com.android.library'
-apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-android-extensions'
-apply from: "${rootDir}/gradle/ktlint.gradle"
-
-android {
-    compileSdkVersion androidVersions.compileSdkVersion
-    buildToolsVersion androidVersions.buildToolsVersion
-
-    defaultConfig {
-        minSdkVersion androidVersions.minSdkVersion
-        targetSdkVersion androidVersions.targetSdkVersion
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles 'proguard-consumer.pro'
-    }
-
-}
-
-dependencies {
-    implementation dependenciesList.kotlinStdLib
-
-    //ktlint
-    ktlint dependenciesList.ktlint
-
-    // Unit testing
-    testImplementation dependenciesList.junit
-    testImplementation dependenciesList.mockito // deprecated
-    testImplementation dependenciesList.mockk
-}
-
-apply from: "${rootDir}/gradle/bintray-publish.gradle"
