--- libnavigation-util/gradle.properties
@@ -1,3 +1,0 @@
-POM_ARTIFACT_ID=utils
-POM_ARTIFACT_TITLE=Mapbox Navigation Utils
-POM_DESCRIPTION=Artifact that provides basic navigation utilities
