--- libnavigation-util/src/main/java/com/mapbox/navigation/utils/span/SpanItem.kt
@@ -1,6 +1,0 @@
-package com.mapbox.navigation.utils.span
-
-interface SpanItem {
-
-    val span: Any
-}
