+++ libtrip-notification/src/main/java/com/mapbox/navigation/trip/notification/MapboxTripNotification.kt
@@ -23,13 +23,13 @@ import com.mapbox.navigation.base.options.NavigationOptions
 import com.mapbox.navigation.base.route.model.LegStepNavigation
 import com.mapbox.navigation.base.route.model.RouteProgressNavigation
 import com.mapbox.navigation.base.trip.TripNotification
+import com.mapbox.navigation.base.utils.END_NAVIGATION_ACTION
+import com.mapbox.navigation.base.utils.NAVIGATION_NOTIFICATION_CHANNEL
+import com.mapbox.navigation.base.utils.NOTIFICATION_CHANNEL
+import com.mapbox.navigation.base.utils.NOTIFICATION_ID
+import com.mapbox.navigation.base.utils.SET_BACKGROUND_COLOR
+import com.mapbox.navigation.base.utils.extensions.ifNonNull
 import com.mapbox.navigation.trip.notification.utils.time.TimeFormatter.formatTime
-import com.mapbox.navigation.utils.END_NAVIGATION_ACTION
-import com.mapbox.navigation.utils.NAVIGATION_NOTIFICATION_CHANNEL
-import com.mapbox.navigation.utils.NOTIFICATION_CHANNEL
-import com.mapbox.navigation.utils.NOTIFICATION_ID
-import com.mapbox.navigation.utils.SET_BACKGROUND_COLOR
-import com.mapbox.navigation.utils.extensions.ifNonNull
 import java.util.Calendar
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.channels.ClosedReceiveChannelException
@@ -118,7 +118,9 @@ class MapboxTripNotification internal constructor(
 
     private fun getNotificationBuilder(): NotificationCompat.Builder {
         val builder =
-            NotificationCompat.Builder(applicationContext, NAVIGATION_NOTIFICATION_CHANNEL)
+            NotificationCompat.Builder(applicationContext,
+                NAVIGATION_NOTIFICATION_CHANNEL
+            )
                 .setCategory(NotificationCompat.CATEGORY_SERVICE)
                 .setPriority(NotificationCompat.PRIORITY_MAX)
                 .setSmallIcon(R.drawable.ic_navigation)
@@ -140,7 +142,8 @@ class MapboxTripNotification internal constructor(
         val collapsedLayoutId = R.id.navigationCollapsedNotificationLayout
         RemoteViews(applicationContext.packageName, collapsedLayout).also { remoteViews ->
             collapsedNotificationRemoteViews = remoteViews
-            remoteViews.setInt(collapsedLayoutId, SET_BACKGROUND_COLOR, backgroundColor)
+            remoteViews.setInt(collapsedLayoutId,
+                SET_BACKGROUND_COLOR, backgroundColor)
         }
 
         val expandedLayout = R.layout.expanded_navigation_notification_layout
@@ -148,7 +151,8 @@ class MapboxTripNotification internal constructor(
         RemoteViews(applicationContext.packageName, expandedLayout).also { remoteViews ->
             expandedNotificationRemoteViews = remoteViews
             remoteViews.setOnClickPendingIntent(R.id.endNavigationBtn, pendingCloseIntent)
-            remoteViews.setInt(expandedLayoutId, SET_BACKGROUND_COLOR, backgroundColor)
+            remoteViews.setInt(expandedLayoutId,
+                SET_BACKGROUND_COLOR, backgroundColor)
         }
     }
