+++ libtrip-notification/src/main/java/com/mapbox/navigation/trip/notification/utils/distance/MapboxDistanceFormatter.kt
@@ -11,9 +11,9 @@ import com.mapbox.navigation.base.typedef.IMPERIAL
 import com.mapbox.navigation.base.typedef.METRIC
 import com.mapbox.navigation.base.typedef.RoundingIncrement
 import com.mapbox.navigation.base.typedef.VoiceUnit
+import com.mapbox.navigation.base.utils.extensions.getUnitTypeForLocale
+import com.mapbox.navigation.base.utils.extensions.inferDeviceLocale
 import com.mapbox.navigation.trip.notification.R
-import com.mapbox.navigation.trip.notification.utils.extension.getUnitTypeForLocale
-import com.mapbox.navigation.utils.extensions.inferDeviceLocale
 import com.mapbox.turf.TurfConstants.UNIT_FEET
 import com.mapbox.turf.TurfConstants.UNIT_KILOMETERS
 import com.mapbox.turf.TurfConstants.UNIT_METERS
